[FEATURE] Output INS to VCF, use args.min_var_length
